fix: adding base classes for accordion, panel, and header
