+++ elements/pfe-accordion/BaseAccordion.scss
@@ -88,106 +88,22 @@ html, body, p, ol, ul, li, dl, dt, dd, blockquote, figure, fieldset, legend, tex
   background-color: transparent;
 }
 
-:host([bordered])::slotted(pfe-accordion-header:first-child) {
+:host([bordered]) ::slotted(pfe-accordion-header:first-child) {
   display: block;
   border-top: 1px solid #d2d2d2;
   border-bottom: 1px solid #d2d2d2;
 }
-:host([bordered])::slotted(pfe-accordion-header:not(:first-child)) {
+:host([bordered]) ::slotted(pfe-accordion-header:not(:first-child)) {
   display: block;
   border-bottom: 1px solid #d2d2d2;
 }
 
-:host([bordered])::slotted(pfe-accordion-header:is([expanded])) {
+:host([bordered]) ::slotted(pfe-accordion-header:is([expanded])) {
   display: block;
   border-bottom: 0px;
 }
 
-:host([bordered])::slotted(pfe-accordion-panel:is([expanded])) {
+:host([bordered]) ::slotted(pfe-accordion-panel:is([expanded])) {
   display: block;
   border-bottom: 1px solid #d2d2d2;
-}
-
-:host([large]) {
-  --pf-c-accordion__toggle--PaddingTop: var(--pf-c-accordion--m-display-lg__toggle--PaddingTop);
-  --pf-c-accordion__toggle--PaddingRight: var(--pf-c-accordion--m-display-lg__toggle--PaddingRight);
-  --pf-c-accordion__toggle--PaddingBottom: var(--pf-c-accordion--m-display-lg__toggle--PaddingBottom);
-  --pf-c-accordion__toggle--PaddingLeft: var(--pf-c-accordion--m-display-lg__toggle--PaddingLeft);
-  --pf-c-accordion__toggle--FontFamily: var(--pf-c-accordion--m-display-lg__toggle--FontFamily);
-  --pf-c-accordion__toggle--FontSize: var(--pf-c-accordion--m-display-lg__toggle--FontSize);
-  --pf-c-accordion__toggle--hover__toggle-text--Color: var(--pf-c-accordion--m-display-lg__toggle--hover__toggle-text--Color);
-  --pf-c-accordion__toggle--active__toggle-text--Color: var(--pf-c-accordion--m-display-lg__toggle--active__toggle-text--Color);
-  --pf-c-accordion__toggle--active__toggle-text--FontWeight: var(--pf-c-accordion--m-display-lg__toggle--active__toggle-text--FontWeight);
-  --pf-c-accordion__toggle--focus__toggle-text--Color: var(--pf-c-accordion--m-display-lg__toggle--focus__toggle-text--Color);
-  --pf-c-accordion__toggle--focus__toggle-text--FontWeight: var(--pf-c-accordion--m-display-lg__toggle--focus__toggle-text--FontWeight);
-  --pf-c-accordion__toggle--m-expanded__toggle-text--Color: var(--pf-c-accordion--m-display-lg__toggle--m-expanded__toggle-text--Color);
-  --pf-c-accordion__toggle--m-expanded__toggle-text--FontWeight: var(--pf-c-accordion--m-display-lg__toggle--m-expanded__toggle-text--FontWeight);
-}
-
-.pf-c-accordion__toggle {
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  width: 100%;
-  padding: var(--pf-c-accordion__toggle--PaddingTop) var(--pf-c-accordion__toggle--PaddingRight) var(--pf-c-accordion__toggle--PaddingBottom) var(--pf-c-accordion__toggle--PaddingLeft);
-  font-family: var(--pf-c-accordion__toggle--FontFamily, inherit);
-  font-size: var(--pf-c-accordion__toggle--FontSize, inherit);
-  border: 0;
-}
-.pf-c-accordion__toggle::after {
-  position: absolute;
-  top: var(--pf-c-accordion__toggle--before--Top);
-  bottom: 0;
-  left: 0;
-  width: var(--pf-c-accordion__toggle--before--Width);
-  content: "";
-  background-color: var(--pf-c-accordion__toggle--before--BackgroundColor);
-}
-.pf-c-accordion__toggle[aria-expanded="true"] {
-  --pf-c-accordion__toggle--before--BackgroundColor: var(--pf-c-accordion__toggle--m-expanded--before--BackgroundColor);
-}
-.pf-c-accordion__toggle[aria-expanded="true"] .pf-c-accordion__toggle-text {
-  font-weight: var(--pf-c-accordion__toggle--m-expanded__toggle-text--FontWeight);
-  color: var(--pf-c-accordion__toggle--m-expanded__toggle-text--Color);
-}
-.pf-c-accordion__toggle[aria-expanded="true"] .pf-c-accordion__toggle-icon {
-  transform: rotate(var(--pf-c-accordion__toggle--m-expanded__toggle-icon--Rotate));
-}
-.pf-c-accordion__toggle:hover {
-  background-color: var(--pf-c-accordion__toggle--hover--BackgroundColor);
-}
-.pf-c-accordion__toggle:hover .pf-c-accordion__toggle-text {
-  color: var(--pf-c-accordion__toggle--hover__toggle-text--Color);
-}
-.pf-c-accordion__toggle:focus {
-  background-color: var(--pf-c-accordion__toggle--focus--BackgroundColor);
-}
-.pf-c-accordion__toggle:focus .pf-c-accordion__toggle-text {
-  font-weight: var(--pf-c-accordion__toggle--focus__toggle-text--FontWeight);
-  color: var(--pf-c-accordion__toggle--focus__toggle-text--Color);
-}
-.pf-c-accordion__toggle:active {
-  background-color: var(--pf-c-accordion__toggle--active--BackgroundColor);
-}
-.pf-c-accordion__toggle:active .pf-c-accordion__toggle-text {
-  font-weight: var(--pf-c-accordion__toggle--active__toggle-text--FontWeight);
-  color: var(--pf-c-accordion__toggle--active__toggle-text--Color);
-}
-
-.pf-c-accordion__toggle-text {
-  overflow: hidden;
-  text-overflow: ellipsis;
-  white-space: nowrap;
-  max-width: var(--pf-c-accordion__toggle-text--MaxWidth);
-}
-
-.pf-c-accordion__toggle-icon {
-  transition: var(--pf-c-accordion__toggle-icon--Transition);
-}
-
-:where(.pf-theme-dark) .pf-c-accordion {
-  --pf-c-accordion__toggle--hover--BackgroundColor: var(--pf-global--BackgroundColor--300, #26292d);
-  --pf-c-accordion__toggle--focus--BackgroundColor: var(--pf-global--BackgroundColor--300, #26292d);
-  --pf-c-accordion__toggle--active--BackgroundColor: var(--pf-global--BackgroundColor--300, #26292d);
 }
