+++ elements/pfe-accordion/BaseAccordionPanel.scss
