+++ elements/pfe-accordion/pfe-accordion-panel.ts
@@ -1,13 +1,6 @@
-import { LitElement, html } from 'lit';
-import { customElement, property } from 'lit/decorators.js';
+import { customElement, } from 'lit/decorators.js';
 
-import { colorContextConsumer, pfelement } from '@patternfly/pfe-core/decorators.js';
-import { getRandomId } from '@patternfly/pfe-core/functions/random.js';
-
-import { PfeCollapsePanel } from '@patternfly/pfe-collapse/pfe-collapse-panel.js';
-
-import style from './pfe-accordion-panel.scss';
-import { ColorTheme } from '@patternfly/pfe-core';
+import { BaseAccordionPanel } from './BaseAccordionPanel.js';
 
 /**
  * Accordion Panel
@@ -15,46 +8,9 @@ import { ColorTheme } from '@patternfly/pfe-core';
  * @slot - Panel content
  * @csspart container - container element for slotted content
  */
-@customElement('pfe-accordion-panel') @pfelement()
-export class PfeAccordionPanel extends LitElement {
+@customElement('pfe-accordion-panel')
+export class PfeAccordionPanel extends BaseAccordionPanel {
   static readonly version = '{{version}}';
-
-  static readonly styles = [
-    ...PfeCollapsePanel.styles,
-    style
-  ];
-
-  @property({ type: String, reflect: true }) boredered?: 'true'|'false';
-
-  @property({ type: Boolean, reflect: true }) expanded = false;
-
-  @property({ type: Boolean, reflect: true }) fixed = false;
-
-  @property({ attribute: 'aria-labelledby', reflect: true }) ariaLabelledby?: string;
-
-  /**
-   * Sets color theme based on parent context
-   */
-  @colorContextConsumer()
-  @property({ reflect: true }) on?: ColorTheme;
-
-  connectedCallback() {
-    super.connectedCallback();
-    this.id ||= getRandomId('pfe-accordion-panel');
-    this.setAttribute('role', 'region');
-  }
-
-  render() {
-    return html`
-      <div tabindex="-1">
-        <div id="container" class="pf-c-accordion__expanded-content" part="container">
-          <div class="pf-c-accordion__expanded-content-body">
-            <slot></slot>
-          </div>
-        </div>
-      </div>
-    `;
-  }
 }
 
 declare global {
